# frontend/src/api/generated/openapi.json
{
  "components": {
    "schemas": {
...
      "CodeCheckRequest": {
        "properties": {
          "code": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Code",
            "type": "string"
          }
        },
        "required": [
          "code"
        ],
        "title": "CodeCheckRequest",
        "type": "object"
      },
      "CodeInfo": {
        "description": "A usable code. `username` is set for reset codes only (the account being reset).",
        "properties": {
          "expires_at": {
            "format": "date-time",
            "title": "Expires At",
            "type": "string"
          },
          "kind": {
            "enum": [
              "invite",
              "reset"
            ],
            "title": "Kind",
            "type": "string"
          },
          "username": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Username"
          }
        },
        "required": [
          "kind",
          "expires_at",
          "username"
        ],
        "title": "CodeInfo",
        "type": "object"
      },
...
      "ErrorCode": {
        "description": "What went wrong; the frontend shows the translation `error.<code>`.",
        "enum": [
          "common.internal",
          "common.not_found",
          "common.method_not_allowed",
          "common.validation",
          "common.rate_limited",
          "common.unauthorized",
          "common.forbidden",
          "common.service_unavailable",
          "auth.invalid_credentials",
          "auth.account_deactivated",
          "auth.token_expired",
          "auth.session_expired",
          "auth.session_revoked",
          "auth.login_required",
          "auth.csrf",
          "auth.code_invalid",
          "auth.password_incorrect",
          "couple.already_in_couple",
          "couple.target_in_couple",
          "couple.request_pending",
          "admin.self_forbidden",
          "admin.last_admin",
          "admin.public_url_missing"
        ],
        "title": "ErrorCode",
        "type": "string"
      },
      "ErrorResponse": {
        "description": "`params` fill placeholders in the translation; `fields` lists rejected request fields.",
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ErrorCode"
          },
          "fields": {
            "items": {
              "$ref": "#/components/schemas/FieldError"
            },
            "title": "Fields",
            "type": "array"
          },
          "params": {
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                },
                {
                  "type": "number"
                },
                {
                  "type": "boolean"
                }
              ]
            },
            "title": "Params",
            "type": "object"
          }
        },
        "required": [
          "code",
          "params",
          "fields"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
      "FieldError": {
        "description": "One rejected request field; `loc` is its path, e.g. `[\"body\", \"name\"]`.",
        "properties": {
          "code": {
            "$ref": "#/components/schemas/FieldErrorCode"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Loc",
            "type": "array"
          }
        },
        "required": [
          "loc",
          "code"
        ],
        "title": "FieldError",
        "type": "object"
      },
      "FieldErrorCode": {
        "description": "Why a single request field was rejected (`fields[].code` in the envelope).",
        "enum": [
          "required",
          "invalid",
          "too_short",
          "too_long",
          "out_of_range",
          "invalid_format",
          "taken",
          "too_common",
          "same_as_username"
        ],
        "title": "FieldErrorCode",
        "type": "string"
      },
      "FilterHidden": {
        "description": "User filter chips this user has switched off, per screen (MEAL-10, UI-02).",
        "properties": {
          "lists": {
            "items": {
              "maxLength": 36,
              "type": "string"
            },
            "maxItems": 500,
            "title": "Lists",
            "type": "array"
          },
          "meals": {
            "items": {
              "maxLength": 36,
              "type": "string"
            },
            "maxItems": 500,
            "title": "Meals",
            "type": "array"
          }
        },
        "required": [
          "meals",
          "lists"
        ],
        "title": "FilterHidden",
        "type": "object"
      },
...
      "JoinRequest": {
        "description": "Registration with an invite code (ACC-05).",
        "properties": {
          "code": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Code",
            "type": "string"
          },
          "display_name": {
            "maxLength": 40,
            "minLength": 1,
            "title": "Display Name",
            "type": "string"
          },
          "language": {
            "enum": [
              "de",
              "en"
            ],
            "title": "Language",
            "type": "string"
          },
          "password": {
            "maxLength": 1024,
            "title": "Password",
            "type": "string"
          },
          "username": {
            "maxLength": 30,
            "minLength": 3,
            "pattern": "^[a-z0-9._-]+$",
            "title": "Username",
            "type": "string"
          }
        },
        "required": [
          "code",
          "username",
          "display_name",
          "password",
          "language"
        ],
        "title": "JoinRequest",
        "type": "object"
      },
...
      "LoginRequest": {
        "properties": {
          "password": {
            "maxLength": 1024,
            "title": "Password",
            "type": "string"
          },
          "username": {
            "maxLength": 100,
            "title": "Username",
            "type": "string"
          }
        },
        "required": [
          "username",
          "password"
        ],
        "title": "LoginRequest",
        "type": "object"
      },
      "LoginResponse": {
        "description": "`access_token` goes into `Authorization: Bearer`; it expires after `expires_in` s.",
        "properties": {
          "access_token": {
            "title": "Access Token",
            "type": "string"
          },
          "expires_in": {
            "title": "Expires In",
            "type": "integer"
          },
          "user": {
            "$ref": "#/components/schemas/Me"
          }
        },
        "required": [
          "access_token",
          "expires_in",
          "user"
        ],
        "title": "LoginResponse",
        "type": "object"
      },
      "Me": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "display_name": {
            "title": "Display Name",
            "type": "string"
          },
          "filter_hidden": {
            "$ref": "#/components/schemas/FilterHidden"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "language": {
            "enum": [
              "de",
              "en"
            ],
            "title": "Language",
            "type": "string"
          },
          "lists_public": {
            "title": "Lists Public",
            "type": "boolean"
          },
          "meals_public": {
            "title": "Meals Public",
            "type": "boolean"
          },
          "role": {
            "enum": [
              "user",
              "admin"
            ],
            "title": "Role",
            "type": "string"
          },
          "username": {
            "title": "Username",
            "type": "string"
          }
        },
        "required": [
          "id",
          "username",
          "display_name",
          "role",
          "language",
          "meals_public",
          "lists_public",
          "filter_hidden",
          "created_at"
        ],
        "title": "Me",
        "type": "object"
      },
...
      "RefreshRequest": {
        "description": "`fork` asks for a new, independent session (first start of the Home Screen app).",
        "properties": {
          "fork": {
            "default": false,
            "title": "Fork",
            "type": "boolean"
          }
        },
        "title": "RefreshRequest",
        "type": "object"
      },
      "ResetRequest": {
        "properties": {
          "code": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Code",
            "type": "string"
          },
          "password": {
            "maxLength": 1024,
            "title": "Password",
            "type": "string"
          }
        },
        "required": [
          "code",
          "password"
        ],
        "title": "ResetRequest",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "description": "The access token from login, refresh, join or reset.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "license": {
      "identifier": "AGPL-3.0-or-later",
      "name": "AGPL-3.0-or-later"
    },
    "title": "MealMate",
    "version": "0.0.0-dev"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/auth/codes/check": {
      "post": {
        "description": "Whether an invite or reset code can be used; does not consume it (ACC-04).",
        "operationId": "check_code",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CodeCheckRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CodeInfo"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Error envelope; `code` names the error"
          }
        },
        "summary": "Check Code",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/api/auth/join": {
      "post": {
        "description": "Register with an invite code and log in (ACC-01, ACC-05, ACC-07).",
        "operationId": "join",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JoinRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Error envelope; `code` names the error"
          }
        },
        "summary": "Join",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/login": {
      "post": {
        "description": "Log in with username and password; sets the refresh cookie (throttled, ACC-11).",
        "operationId": "login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Error envelope; `code` names the error"
          }
        },
        "summary": "Login",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/logout": {
      "post": {
        "description": "Log out this device: revoke the cookie's session and remove the cookie.",
        "operationId": "logout",
        "parameters": [
          {
            "in": "cookie",
            "name": "mm_refresh",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Mm Refresh"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Error envelope; `code` names the error"
          }
        },
        "summary": "Logout",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/logout-all": {
      "post": {
        "description": "Log out every device of the user, this one included (ACC-09).",
        "operationId": "logout_all",
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Error envelope; `code` names the error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Logout All",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/refresh": {
      "post": {
        "description": "A new access token (and rotated cookie) for the refresh cookie.\n\nNeeds the header `X-MealMate-Client: web` (CSRF guard). `{\"fork\": true}` starts a new,\nindependent session instead (first start of the Home Screen app).",
        "operationId": "refresh",
        "parameters": [
          {
            "in": "header",
            "name": "X-MealMate-Client",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Mealmate-Client"
            }
          },
          {
            "in": "cookie",
            "name": "mm_refresh",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Mm Refresh"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/RefreshRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Error envelope; `code` names the error"
          }
        },
        "summary": "Refresh",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/reset": {
      "post": {
        "description": "Set a new password with a reset code; ends all other sessions and logs in (ACC-10).",
        "operationId": "reset_password",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Error envelope; `code` names the error"
          }
        },
        "summary": "Reset Password",
        "tags": [
          "auth"
        ]
      }
    },
...
  }
}